Clean up + Moved function around
